Move Transporter requests to a Transporter namespace
