--- src/Requests/Messaging/BulkSmsRequest.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace SamuelMwangiW\Africastalking\Requests\Messaging;
-
-use SamuelMwangiW\Africastalking\Requests\AfricastalkingRequest;
-
-class BulkSmsRequest extends AfricastalkingRequest
-{
-    protected string $path = 'messaging';
-}
