+++ src/Transporter/Requests/Messaging/BulkSmsRequest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace SamuelMwangiW\Africastalking\Transporter\Requests\Messaging;
+
+use SamuelMwangiW\Africastalking\Transporter\Requests\AfricastalkingRequest;
+
+class BulkSmsRequest extends AfricastalkingRequest
+{
+    protected string $path = 'messaging';
+}
