+++ src/Transporter/Requests/Messaging/PremiumSmsRequest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace SamuelMwangiW\Africastalking\Requests\Messaging;
+namespace SamuelMwangiW\Africastalking\Transporter\Requests\Messaging;
 
-use SamuelMwangiW\Africastalking\Requests\AfricastalkingRequest;
+use SamuelMwangiW\Africastalking\Transporter\Requests\AfricastalkingRequest;
 
 class PremiumSmsRequest extends AfricastalkingRequest
 {
